RPM: Make protocol installation optional. (#44)
RPM occupies a SMD (shared memory device) channel, hence it is necessary to
release RPM channel before transferring control to the OS. However, what we
can do is limited in ExitBootServices callback.

We used to workaround the problem using a while-loop combined with thread
sleep, but the mitigation is problematic under the context of uni-core
environment. Since RPM usage is limited in current UEFI implementation,
we will release the RPM channel as soon as possible and using the
non-blocking signaling (semaphore) mechanism.